Added util/Models.ts
